moved component to multichain folder
