+++ ui/components/multichain/multichain-connected-site-menu/multichain-connected-site-menu.stories.js
@@ -1,11 +1,11 @@
 import React from 'react';
 import { Provider } from 'react-redux';
-import testData from '../../../.storybook/test-data';
-import configureStore from '../../store/store';
+import testData from '../../../../.storybook/test-data';
+import configureStore from '../../../store/store';
 import { MultichainConnectedSiteMenu } from './multichain-connected-site-menu';
 
 export default {
-  title: 'Components/UI/MultichainConnectedSiteMenu',
+  title: 'Components/Multichain/MultichainConnectedSiteMenu',
 
   component: MultichainConnectedSiteMenu,
 };
